gui-main/requirements.txt
# Spine Ultrasound Platform - Ubuntu 22.04 mainline dependencies
PySide6>=6.7
PyYAML>=6.0
pytest>=8.0
pyqtgraph>=0.13.0
numpy>=1.24.0
protobuf>=3.20.3,<4
fastapi>=0.100.0
uvicorn>=0.23.0
loguru>=0.7.0
httpx>=0.27.0
